+++ administracion/urls.py
@@ -2,7 +2,7 @@ from django.conf.urls import url
 from .views import admin_users_dashboard, \
                    admin_users_create, admin_users_edit, admin_users_edit_form, \
                    admin_users_delete_modal, admin_users_delete, list_studies, \
-                   focus_mode, reject_study, search_students, detail_student
+                   focus_mode, search_students, detail_student
 
 app_name = 'administracion'
 
@@ -16,7 +16,6 @@ urlpatterns = [
     url(r'^usuarios/', admin_users_dashboard, name='users'),
     url(r'^principal/(?P<status_study>[\w\-]+)/$', list_studies, name='main_estudios'),
     url(r'^principal/(?P<study_id>[\w\-]+)/detalle', focus_mode, name='focus_mode'),
-    url(r'^estudio/rechazar/$', reject_study, name='reject_study'),
     url(r'^busqueda/', search_students, name='search_students'),
     url(r'^detalle-alumno/(?P<id_alumno>[0-9]+)', detail_student, name='detail_student'),
 ]
