+++ jp2_online/templates/estudios_socioeconomicos/focus_mode_seccion_revision.html
@@ -1,0 +1,59 @@
+{% load permissions_extras %}
+
+{% if request.user|has_group:'Administrador' %}
+<div class="row">
+    <div id="accept_section" class="col-sm-12 col-md-6 col-lg-6">
+        <h4 class="center">Aceptar Estudio</h4>
+        <br><br>
+        <p class="center">
+            Una vez que se acepte el estudio, este ya no 
+            estara disponible para el capturista. Posteriormente, 
+            en la sección de
+            estudios aprobados se les podra asignar la beca a los alumnos de este estudio.
+        </p>
+        <br><br>
+        <form method="post" action="{% url 'estudios_socioeconomicos:accept_study' estudio.id %}">
+            {% csrf_token %}
+            <button class="btn btn-success center">
+                Aceptar
+            </button>
+        </form>
+    </div>
+
+    <div class="col-sm-12 col-md-6 col-lg-6">
+        <form method="post" action="{% url 'estudios_socioeconomicos:reject_study' %}">
+            <h4 class="center">Rechazar Estudio</h4>
+            <br><br>
+            <p class="center">
+                Al rechazar el estudio se debera de dejar retroalimentación
+                para que el capturista pueda hacer correcciones y volverlo a 
+                subir.
+            </p>
+            <br><br>
+            {% csrf_token %}
+            {{feedback_form}}
+            
+            <br><br>
+
+            
+            <button class="btn btn-danger center">
+                Rechazar
+            </button>
+        </form>  
+    </div>
+
+</div>
+{% endif %}
+
+<style>
+    #accept_section {
+        border-right: solid;
+        border-width: 1px;
+    }
+
+    .center{
+        display: table;
+        margin: 0 auto;
+    }
+
+</style>
